# api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Starvis API",
    "version": "1.0.0",
    "description": "Star Citizen Ships, Components, Items & Game Data API.\n\nData sources:\n- **P4K game data** (ships, components, items, commodities, shops, missions, mining, crafting, locations)\n- **RSI Website** (ship matrix, galactapedia, comm-links, starmap)\n\n### Authentication\nMost endpoints are public. Two authentication methods are supported:\n- **Bearer JWT** (`Authorization: Bearer <token>`): obtained via `POST /auth/login` or `POST /auth/api-token`. Required for `/auth/me`, beta-only tools, and admin routes.\n- **API Key** (`X-API-Key` header): for admin endpoints only.\n\n### User Roles\n| Role | Description |\n|---|---|\n| `user` | Standard account — access to all public data |\n| `beta_tester` | Early access to tools in active development (Loadout Manager, FPS Calculator, Mining Calculator, Trade Calculator, Crafting Calculator) |\n| `admin` | Full access including user management and admin API |\n\nRoles are assigned by an admin via `PUT /admin/users/{id}/role`.\n\n### Pagination\nAll list endpoints support `page` and `limit` query parameters and return `{ data, total, page, limit, pages }`.\n\n### Filters\nEvery domain has a `/filters` endpoint returning `{ filters: { fieldName: [{ value, label, count? }] } }`.\n\n### Game environments\nUse the `env` query parameter to select the game environment: `live` (default), `ptu`, `eptu`, `custom`."
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
...
      "search": {
        "name": "search",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Full-text search query"
      },
...
      "sort": {
        "name": "sort",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Field to sort by (supports dot-notation for JSON fields, e.g. `game_data.scm_speed`)"
      },
...
    },
...
  },
...
  "paths": {
    "/api/v1/ships": {
      "get": {
        "tags": ["Ships"],
        "summary": "List space ships",
        "operationId": "getShips",
        "description": "Returns space ships only (`vehicle_category = ship`). Use `/ground-vehicles` or `/gravlev` for other vehicle types.",
        "parameters": [
          {
            "$ref": "#/components/parameters/search"
          },
          {
            "name": "manufacturer",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Manufacturer code (e.g. AEGS, RSI)"
          },
          {
            "name": "career",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Career (Combat, Transport, …)"
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Role (Light Fighter, Medium Freight, …)"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["flight-ready", "in-concept", "in-production", "in-game-only"]
            }
          },
          {
            "name": "variant_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated relations: manufacturer, paints, similar, matrix"
          },
          {
            "$ref": "#/components/parameters/sort"
          },
          {
            "$ref": "#/components/parameters/order"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/format"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated ship list",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        }
                      }
                    },
                    {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Ship"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
    "/api/v1/ground-vehicles": {
      "get": {
        "tags": ["Ground Vehicles"],
        "summary": "List ground vehicles",
        "operationId": "getGroundVehicles",
        "description": "Returns ground vehicles only (`vehicle_category = ground`): buggies, tanks, rovers, etc.",
        "parameters": [
          {
            "$ref": "#/components/parameters/search"
          },
          {
            "name": "manufacturer",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "career",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/sort"
          },
          {
            "$ref": "#/components/parameters/order"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/env"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated ground vehicle list",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Ship"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
    "/api/v1/gravlev": {
      "get": {
        "tags": ["Grav-lev"],
        "summary": "List grav-lev vehicles",
        "operationId": "getGravlev",
        "description": "Returns grav-lev vehicles only (`vehicle_category = gravlev`): hovercycles, nox bikes, etc.",
        "parameters": [
          {
            "$ref": "#/components/parameters/search"
          },
          {
            "name": "manufacturer",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/sort"
          },
          {
            "$ref": "#/components/parameters/order"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/env"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated grav-lev vehicle list",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Ship"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
    "/api/v1/ships/search": {
      "get": {
        "tags": ["Ships"],
        "summary": "Autocomplete ship search",
        "operationId": "searchShipsAutocomplete",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching ships (lightweight)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Ship"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Search query too short"
          }
        }
      }
    },
...
    "/api/v1/ships/manufacturers": {
      "get": {
        "tags": ["Ships"],
        "summary": "Get manufacturers with ships",
        "operationId": "getShipManufacturers",
        "parameters": [
          {
            "$ref": "#/components/parameters/format"
          }
        ],
        "responses": {
          "200": {
            "description": "Manufacturer list with ship count"
          }
        }
      }
    },
    "/api/v1/ships/{uuid}": {
      "get": {
        "tags": ["Ships"],
        "summary": "Get ship by UUID or class_name",
        "operationId": "getShipByUuid",
        "description": "Returns detailed ship information with optional related data. Use `include` parameter to fetch manufacturer details, paints, similar ships, and RSI Ship Matrix entry.",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Ship UUID or class_name"
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated relations: manufacturer, paints, similar, matrix. Similar ships have same career/role; matrix includes official RSI specs."
          }
        ],
        "responses": {
          "200": {
            "description": "Ship details with optional relations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ShipDetail"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/components": {
      "get": {
        "tags": ["Components"],
        "summary": "List components",
        "operationId": "getComponents",
        "parameters": [
          {
            "$ref": "#/components/parameters/search"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Component type (WeaponGun, Shield, QuantumDrive, …)"
          },
          {
            "name": "sub_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "grade",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "manufacturer",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/sort"
          },
          {
            "$ref": "#/components/parameters/order"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/format"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated component list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Component"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pages": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/components/{uuid}": {
      "get": {
        "tags": ["Components"],
        "summary": "Get component by UUID or class_name",
        "operationId": "getComponentByUuid",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated: manufacturer"
          }
        ],
        "responses": {
          "200": {
            "description": "Component details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Component"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/items": {
      "get": {
        "tags": ["Items"],
        "summary": "List FPS items",
        "operationId": "getItems",
        "parameters": [
          {
            "$ref": "#/components/parameters/search"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Item type (Weapon_Personal, Char_Armor_Helmet, …)"
          },
          {
            "name": "sub_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "manufacturer",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/sort"
          },
          {
            "$ref": "#/components/parameters/order"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/format"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated item list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Item"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/items/manufacturers": {
      "get": {
        "tags": ["Items"],
        "summary": "List manufacturers for a given item type",
        "operationId": "getItemManufacturers",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/env"
          }
        ],
        "responses": {
          "200": {
            "description": "Manufacturers list"
          }
        }
      }
    },
...
    "/api/v1/items/{uuid}": {
      "get": {
        "tags": ["Items"],
        "summary": "Get item by UUID or class_name",
        "operationId": "getItemByUuid",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated: manufacturer"
          }
        ],
        "responses": {
          "200": {
            "description": "Item details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Item"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/commodities": {
      "get": {
        "tags": ["Commodities"],
        "summary": "List commodities",
        "operationId": "getCommodities",
        "parameters": [
          {
            "$ref": "#/components/parameters/search"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_illegal",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "$ref": "#/components/parameters/sort"
          },
          {
            "$ref": "#/components/parameters/order"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/format"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated commodity list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Commodity"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/commodities/{uuid}": {
      "get": {
        "tags": ["Commodities"],
        "summary": "Get commodity by UUID or class_name",
        "operationId": "getCommodityByUuid",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commodity details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Commodity"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/paints": {
      "get": {
        "tags": ["Paints"],
        "summary": "List all ship paints",
        "operationId": "getPaints",
        "parameters": [
          {
            "name": "ship_uuid",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by ship UUID"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/format"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated paint list"
          }
        }
      }
    },
...
  }
}